Fix OOM: chunked atom-centered feature extraction
The previous version stored all 134k molecule activations in memory
simultaneously (~12GB), then OOM-killed during np.hstack. Now
processes molecules in 5000-molecule chunks: compute vertex
activations, extract ADE features, write to pre-allocated float32
output matrix, free intermediates. Peak memory ~4GB instead of 20GB.

Also caches X^T X in the ridge solver so the alpha sweep doesn't
recompute the dominant O(N*D^2) product per alpha.

QM9 atom-centered run in progress — waiting for results.

+++ train_qm9_local.py
@@ -135,9 +135,9 @@ def main():
     print(f"  NaN: {n_nan}, Inf: {n_inf}, dead: {dead}/{X_all.shape[1]}")
     X_all = np.nan_to_num(X_all)
 
-    # Split and standardize
-    X_train = X_all[train_idx]
-    X_test = X_all[test_idx]
+    # Split and standardize (upcast to float64 for ridge numerical stability)
+    X_train = X_all[train_idx].astype(np.float64)
+    X_test = X_all[test_idx].astype(np.float64)
     del X_all; gc.collect()
 
     mu = X_train.mean(axis=0)
